# yuanzi-backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "小园子母婴记录应用后端 API (基于 PRD-V1.1)",
        "title": "Yuanzi API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/ai/chat": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "发送问题给 AI 助手获取回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI 问答",
                "parameters": [
                    {
                        "description": "问答请求",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AIChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AIChatResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/ai/quota": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取用户当日 AI 使用配额情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "获取 AI 配额",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AIQuotaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/ai/speech/recognize": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "将语音转为文字",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "语音识别",
                "parameters": [
                    {
                        "type": "file",
                        "description": "音频文件",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.SpeechRecognizeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/device/register": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "注册设备 token，用于推送通知",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送"
                ],
                "summary": "注册推送设备",
                "parameters": [
                    {
                        "description": "设备信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.AIChatRequest": {
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "baby_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ChatHistory"
                    }
                },
                "question": {
                    "type": "string",
                    "example": "宝宝三个月应该喝多少奶"
                }
            }
        },
        "handler.AIChatResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "remaining_quota": {
                    "type": "integer"
                },
                "tokens_used": {
                    "type": "integer"
                }
            }
        },
        "handler.AIQuotaResponse": {
            "type": "object",
            "properties": {
                "ai_chat": {
                    "$ref": "#/definitions/handler.QuotaDetail"
                },
                "speech": {
                    "$ref": "#/definitions/handler.QuotaDetail"
                }
            }
        },
        "handler.ChatHistory": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "之前的问题"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "handler.QuotaDetail": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "remaining": {
                    "type": "integer"
                },
                "used": {
                    "type": "integer"
                }
            }
        },
...
        "handler.RegisterDeviceRequest": {
            "type": "object",
            "required": [
                "device_token",
                "platform"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "device_token": {
                    "type": "string"
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "ios",
                        "android"
                    ]
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "handler.SpeechRecognizeResponse": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "remaining_quota": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}